pyproject.toml
[project]
name = "echo-pipecat-agent"
version = "0.2.0"
description = "Pipecat voice agent with Gemini STT/LLM and Echo TTS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[daily,google,silero,webrtc,runner]",
    "websockets",
    "numpy",
    "aiohttp",
    "loguru",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "pymongo",
    "google-cloud-storage",
    "transformers>=4.57.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# This is a script-based project, not a package
# Just include the Python files
only-include = [
    "bot.py",
    "custom_echo_tts_service.py",
    "custom_vibevoice_tts_service.py",
    "test_local.py",
]
